+++ Calibrator.hpp
@@ -23,10 +23,10 @@ class Calibrator{
 		EventBuilder *myEventBuilder;
 		std::deque <ADCDataItem> eventList;		//Event list read in from buffer
 
-		const unsigned int feeChannelOrder[64]={62, 63, 59, 60, 61, 56, 57, 58, 52, 53, 54, 55, 49, 50, 51, 45,
-												46, 47, 48, 42, 43, 44, 38, 39, 40, 41, 35, 36, 37, 31, 32, 33,
-			    								34, 28, 29, 30, 24, 25, 26, 27, 21, 22, 23, 17, 18, 19, 20, 14,
-												15, 16, 10, 11, 12,  7,  3,  0,  8,  4,  1,  9,  5,  2, 13,  6};
+		const double feeChannelOrder[64]={62., 63., 59., 60., 61., 56., 57., 58., 52., 53., 54., 55., 49., 50., 51., 45.,
+												46., 47., 48., 42., 43., 44., 38., 39., 40., 41., 35., 36., 37., 31., 32., 33.,
+												34., 28., 29., 30., 24., 25., 26., 27., 21., 22., 23., 17., 18., 19., 20., 14.,
+												15., 16., 10., 11., 12.,  7.,  3.,  0.,  8.,  4.,  1.,  9.,  5.,  2., 13.,  6.};
 
 		const double adcZero = 32768; 				//2**15
 
@@ -45,7 +45,7 @@ class Calibrator{
 		void SetGeometry(ADCDataItem & adcDataItemIn, CalibratedADCDataItem & calibratedItemOut);
 		void CalibrateData(ADCDataItem & adcDataItemIn, CalibratedADCDataItem & calibratedItemOut);
 		bool CalibrateEnergy(ADCDataItem & adcDataItemIn, CalibratedADCDataItem & calibratedItemOut);
-		unsigned int GetOrder(int channelID) const;
+		double GetOrder(int channelID) const;
 
 	public:
 		Calibrator();
